requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
requests==2.31.0
urllib3==2.0.3
aiohttp==3.7.4
aiohttp==3.7.4
aiofiles==0.7.0
fastapi==0.70.0
uvicorn==0.15.0
psycopg2-binary==2.9.1
